chore: patch cypress deps
