--- src/test/cypress/cypress.json
@@ -1,25 +1,0 @@
-{
-  "$schema": "https://on.cypress.io/cypress.schema.json",
-  "video": false,
-  "fixturesFolder": "e2e/fixtures",
-  "integrationFolder": "e2e/integration",
-  "supportFile": "e2e/support/index.js",
-  "pluginsFile": "e2e/plugins/index.js",
-  "downloadsFolder": "downloads",
-  "screenshotOnRunFailure": true,
-  "screenshotsFolder": "screenshots",
-  "trashAssetsBeforeRuns": true,
-  "videosFolder": "videos",
-  "viewportHeight": 768,
-  "viewportWidth": 1536,
-  "requestTimeout": 10000,
-  "defaultCommandTimeout": 8000,
-  "reporter": "junit",
-  "reporterOptions": {
-    "mochaFile": "reports/result-[hash].xml"
-  },
-  "retries": {
-    "runMode": 1,
-    "openMode": 0
-  }
-}
