+++ src/test/cypress/package.json
@@ -21,10 +21,10 @@
   },
   "devDependencies": {
     "axe-core": "^4.4.2",
-    "cypress": "^9.6.1",
+    "cypress": "^10.0.1",
     "cypress-axe": "^0.14.0",
     "cypress-plugin-tab": "^1.0.5",
-    "eslint": "^8.15.0",
+    "eslint": "^8.16.0",
     "eslint-plugin-cypress": "^2.12.1",
     "fs-extra": "^10.1.0",
     "path": "^0.12.7",
